yaw semantics, altitude floor, parse-vs-zmq split, abort idempotent, battery float
Major:
- set_velocity_ned advertised a "Yaw rate (deg/s)" arg called
  yaw_rate_deg_s but stuffed it into Offboard::VelocityNedYaw.yaw_deg,
  which MAVSDK documents as ABSOLUTE heading (0=North, CW positive).
  yaw rate lives on a different type (VelocityBodyYawspeed). Same arg
  shape on goto_ned: optional_double defaulted absent values to 0.0,
  so every call without an explicit yaw silently snapped the drone to
  face north. Both tools now: rename the arg to yaw_deg, document it
  as absolute heading, and when absent forward the current telemetry
  yaw (no MAVSDK sentinel for "leave heading alone").
- Safety gate accepted negative takeoff altitudes and goto_ned d > 0
  (5 m underground): only the upper bound was checked. Added explicit
  altitude_non_positive (takeoff) and altitude_below_home (goto_ned)
  rejections so the documented envelope catches what PX4 might or
  might not.

Medium:
- _dispatch_tool_call's wire-error catch flipped _bridge_offline_flag
  on ValueError/TypeError, which fire when the bridge replies with
  garbled JSON or a non-dict. Bridge replied — so it's alive. Flagging
  offline made every subsequent dispatch in the mission short-circuit
  at the gate. Split the catch: zmq.ZMQError still flags offline (real
  transport failure); ValueError/TypeError reset REQ state but don't.
- A second voice abort during a latched abort hit the safety gate's
  abort_active denial and surfaced as abort_failed even though the
  original abort was still descending. abort joined kill/land/disarm
  in the emergency-followup carveout — Action::land() on a landing
  drone is a no-op, so a double-press is harmless and idempotent.
- services_panel battery_pct used get<int>(); a float (PX4 has been
  observed reporting fractional percentages on some FC builds) would
  throw type_error and the outer catch wiped the entire telemetry to
  "parse error". Now get<double>() then cast.

+++ src/services_panel.cpp
@@ -81,7 +81,8 @@ std::string summarize_telemetry(const std::string& header) {
             ss << " / " << j["rel_alt_m"].get<double>() << " m";
         }
         if (j.contains("battery_pct") && !j["battery_pct"].is_null()) {
-            ss << " / " << j["battery_pct"].get<int>() << "%";
+            // get<double>() tolerates both int and float without throwing.
+            ss << " / " << static_cast<int>(j["battery_pct"].get<double>()) << "%";
         }
         return ss.str();
     } catch (...) {
@@ -107,7 +108,7 @@ void parse_telemetry(const std::string& header, TelemetrySnapshot& out) {
         const bool has_pct = j.contains("battery_pct") && !j["battery_pct"].is_null();
         const bool has_v   = j.contains("battery_v")   && !j["battery_v"].is_null();
         out.has_battery = has_pct || has_v;
-        if (has_pct) out.battery_pct = j["battery_pct"].get<int>();
+        if (has_pct) out.battery_pct = static_cast<int>(j["battery_pct"].get<double>());
         if (has_v)   out.battery_v   = j["battery_v"].get<float>();
 
         if (j.contains("lat") && !j["lat"].is_null() &&
